% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_miles.R
\name{cumulative_miles}
\alias{cumulative_miles}
\title{Calculate cumulative miles}
\usage{
cumulative_miles(d, miles_per_card = 5)
}
\arguments{
\item{d}{A \code{data.frame} contain all the cards with day information.}

\item{miles_per_card}{An integer specifying how many miles a card is worth.}
}
\value{
A \code{data.frame} containing the number of miles run each week and 
the cumulative miles per week.
}
\description{
Calculates the cumulative miles by week.
}
\examples{
\dontrun{
read_cards() \%>\% cumulative_miles
}
}

