% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_miles_by_id.R
\name{cumulative_miles_by_id}
\alias{cumulative_miles_by_id}
\title{Calculate cumulative miles for each individual by week}
\usage{
cumulative_miles_by_id(d, miles_per_card = 5)
}
\arguments{
\item{d}{A \code{data.frame} containing all the card and day information.}

\item{miles_per_card}{An integer specifying how many miles a card is worth.}
}
\value{
A \code{data.frame} with the cumulative miles for each individual by week.
}
\description{
For each individual, calculate the miles run that week and the 
cumulative miles run up to and including that week.
}
\examples{
\dontrun{
read_cards() \%>\% cumulative_miles_by_id
}
}

