% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_name.R
\name{find_name}
\alias{find_name}
\title{Find a student's name or ID}
\usage{
find_name(ids, school = "Fellows")
}
\arguments{
\item{ids}{A \code{data.frame} containing the student}

\item{school}{A string containing the name of the school}
}
\description{
Launch a shiny app to help find a student's name or ID
}
\examples{
\dontrun{
find_name()
}
}

