% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestones.R
\name{milestones}
\alias{milestones}
\title{Calculate milestones}
\usage{
milestones(d)
}
\arguments{
\item{d}{A \code{data.frame} containing all card and day information.}
}
\value{
A \code{data_frame} with the number of students hitting a particular milestone.
}
\description{
This will calculate, by week, the number of students hitting a particular
milestone (every 5 miles) and the cumulative number of students who have hit
that milestone as of that week.
}
\examples{
\dontrun{
read_cards() \%>\% milestones
}
}

