% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_read_csv.R
\name{my_read_csv}
\alias{my_read_csv}
\title{Read csv file and extract date column from filename}
\usage{
my_read_csv(f)
}
\arguments{
\item{f}{A YYYYMMDD.csv file}
}
\value{
A \code{data.frame} containing the information from the csv file as well as a date column from the filename itself.
}
\description{
Read csv file whose filename is the date in YYYYMMDD format and 
add this date as a column in the returned data.frame.
}
\examples{
\dontrun{
my_read_csv('20160416.csv')
}
}

