% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cards.R
\name{read_cards}
\alias{read_cards}
\title{Read all cards and merge with days information}
\usage{
read_cards(cards_path = "data/cards/", days_file = "data/days.csv")
}
\arguments{
\item{cards_path}{A directory contain the card files.}

\item{days_file}{A directory contain the day information.}
}
\value{
A \code{data.frame} of all cards merged with day information.
}
\description{
Read all card files. The filenames should be YYYYMMDD.csv and 
should have a single column labeled 'id'.
}
\examples{
\dontrun{
read_cards()
}
}

