% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Get mileage club summary}
\usage{
summary(cards, school = "Fellows")
}
\arguments{
\item{cards}{A \code{data.frame} containing all the cards with the weeks that they were completed.}

\item{school}{A string containing the name of the school}
}
\description{
Get the mileage club summary as a shiny app which reports the miles run each week,
the cumulative miles run, and some other information.
}
\examples{
\dontrun{
summary()
}
}

