% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Get mileage club summary}
\usage{
summary(cards, ids, school = "Fellows")
}
\arguments{
\item{cards}{A \code{data.frame} containing all the cards with the weeks that they were completed.}

\item{ids}{A \code{data.frame} containing the student}

\item{school}{A string containing the name of the school}
}
\description{
Get the mileage club summary as a shiny app which reports the cumulative miles, milestones, and students
reaching milestones.
}
\examples{
\dontrun{
cards = read_cards()
ids = read_csv('data/id.csv')
summary(cards,ids)
}
}

